requirements.txt
tornado>=6.3.0
torch>=2.2.0
torchvision>=0.17.0
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0
